requirements.txt
torch==2.6.0
numpy
wandb
ale-py
gymnasium
pygame
opencv-python
PyYAML